requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
asyncpg==0.31.0
certifi==2026.2.25
click==8.3.2
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.135.3
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
psycopg2-binary==2.9.11
pydantic==2.13.0
pydantic-settings==2.13.1
pydantic_core==2.46.0
python-dotenv==1.2.2
SQLAlchemy==2.0.35
uuid6==2024.7.10
PyYAML==6.0.3
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uuid6==2025.0.1
uvicorn==0.44.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0
